+++ subgraphs/ethereum/tests/liquidityCurve.test.ts
@@ -1,5 +1,5 @@
-import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
-import { assert, beforeEach, clearStore, createMockedFunction, describe, test } from "matchstick-as/assembly/index";
+import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
+import { assert, beforeEach, clearStore, createMockedFunction, describe, test, log } from "matchstick-as/assembly/index";
 
 import { TokenCategoryStable } from "../../shared/src/contracts/TokenDefinition";
 import { toBigInt } from "../../shared/src/utils/Decimals";
@@ -53,18 +53,22 @@ import {
   mockUniswapV2PairsZero,
   mockUniswapV3PairsZero,
   mockUsdOhmV2Rate,
-  OHM_USD_RESERVE_BLOCK,
   OHM_V2_DECIMALS,
 } from "./pairHelper";
 import { mockWalletBalance, mockZeroWalletBalances } from "./walletHelper";
 import { getWalletAddressesForContract } from "../src/utils/ProtocolAddresses";
+import { mockTreasuryAddressNull } from "./bophadesHelper";
 
 const PAIR_CURVE_OHM_ETH_TOTAL_SUPPLY = BigDecimal.fromString("100");
 const TIMESTAMP = BigInt.fromString("1");
+const BLOCK_NUMBER: BigInt = BigInt.fromString("14000000");
 
 beforeEach(() => {
   log.debug("beforeEach: Clearing store", []);
   clearStore();
+
+  // Do at the start, as it can be used by mock functions
+  mockTreasuryAddressNull();
 
   mockBalancerVaultZero();
   mockUniswapV2PairsZero();
@@ -98,7 +102,7 @@ describe("Token Quantity", () => {
     const totalTokenQuantity = getCurvePairTotalTokenQuantity(
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V2,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     assert.stringEquals(totalTokenQuantity.toString(), ohmReserves.toString());
@@ -134,9 +138,9 @@ describe("Token Quantity", () => {
 
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockWalletBalance(
       ERC20_CRV_OHMETH,
       TREASURY_ADDRESS_V3,
@@ -147,7 +151,7 @@ describe("Token Quantity", () => {
       TIMESTAMP,
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V2,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     // Should be empty records due to starting block
@@ -191,9 +195,9 @@ describe("Token Quantity", () => {
 
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockWalletBalance(
       ERC20_CRV_OHMETH,
       TREASURY_ADDRESS_V3,
@@ -204,7 +208,7 @@ describe("Token Quantity", () => {
       TIMESTAMP,
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V2,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     // Should be empty records due to starting block
@@ -241,9 +245,9 @@ describe("Token Quantity", () => {
 
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockFraxLockedBalanceZero();
     mockWalletBalance(
       ERC20_CRV_OHMETH,
@@ -258,7 +262,7 @@ describe("Token Quantity", () => {
       TIMESTAMP,
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V2,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     // Balance = value as the unit rate is 1
@@ -296,9 +300,9 @@ describe("Token Quantity", () => {
 
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockWalletBalance(
       ERC20_CRV_OHMETH,
       TREASURY_ADDRESS_V3,
@@ -309,7 +313,7 @@ describe("Token Quantity", () => {
       TIMESTAMP,
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V1,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     // Should be empty records due to 0 balance of OHM V1
@@ -346,9 +350,9 @@ describe("Token Quantity", () => {
 
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockFraxLockedBalanceZero();
     mockConvexStakedBalance(
       ERC20_CVX_OHMETH,
@@ -364,7 +368,7 @@ describe("Token Quantity", () => {
       TIMESTAMP,
       PAIR_CURVE_OHM_ETH,
       ERC20_OHM_V2,
-      OHM_USD_RESERVE_BLOCK,
+      BLOCK_NUMBER,
     );
 
     // Balance = value as the unit rate is 1
@@ -397,7 +401,7 @@ describe("Pair Value", () => {
     const expectedValue = getPairValue(ohmBalance, wethBalance, getOhmUsdRate(), getEthUsdRate());
     assert.stringEquals(
       expectedValue.truncate(4).toString(),
-      getCurvePairTotalValue(PAIR_CURVE_OHM_ETH, false, OHM_USD_RESERVE_BLOCK).truncate(4).toString(),
+      getCurvePairTotalValue(PAIR_CURVE_OHM_ETH, false, BLOCK_NUMBER).truncate(4).toString(),
     );
   });
 
@@ -424,7 +428,7 @@ describe("Pair Value", () => {
     const expectedValue = wethBalance.times(getEthUsdRate());
     assert.stringEquals(
       expectedValue.truncate(4).toString(),
-      getCurvePairTotalValue(PAIR_CURVE_OHM_ETH, true, OHM_USD_RESERVE_BLOCK).truncate(4).toString(),
+      getCurvePairTotalValue(PAIR_CURVE_OHM_ETH, true, BLOCK_NUMBER).truncate(4).toString(),
     );
   });
 
@@ -435,7 +439,7 @@ describe("Pair Value", () => {
     mockFraxLockedBalanceZero();
     mockZeroWalletBalances(
       ERC20_BALANCER_OHM_DAI_WETH,
-      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID),
+      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID, BLOCK_NUMBER),
     );
 
     mockEthUsdRate();
@@ -465,16 +469,16 @@ describe("Pair Value", () => {
     );
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockWalletBalance(
       ERC20_CRV_OHMETH,
       TREASURY_ADDRESS_V3,
       toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
     );
 
-    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, OHM_USD_RESERVE_BLOCK);
+    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, BLOCK_NUMBER);
 
     const expectedNonOhmTotalValue = wethReserves.times(getEthUsdRate());
     const expectedTotalValue = getPairValue(
@@ -502,7 +506,7 @@ describe("Pair Value", () => {
     mockFraxLockedBalanceZero();
     mockZeroWalletBalances(
       ERC20_BALANCER_OHM_DAI_WETH,
-      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID),
+      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID, BLOCK_NUMBER),
     );
 
     mockEthUsdRate();
@@ -532,16 +536,16 @@ describe("Pair Value", () => {
     );
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockWalletBalance(
       ERC20_CRV_OHMETH,
       TREASURY_ADDRESS_V3,
       toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
     );
 
-    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, OHM_USD_RESERVE_BLOCK);
+    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, BLOCK_NUMBER);
 
     assert.i32Equals(0, records.length);
   });
@@ -553,7 +557,7 @@ describe("Pair Value", () => {
     mockFraxLockedBalanceZero();
     mockZeroWalletBalances(
       ERC20_BALANCER_OHM_DAI_WETH,
-      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID),
+      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID, BLOCK_NUMBER),
     );
 
     // Mock price lookup
@@ -584,9 +588,9 @@ describe("Pair Value", () => {
     );
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
+    mockZeroWalletBalances(ERC20_CRV_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_OHMETH, getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
     mockConvexStakedBalance(
       ERC20_CVX_OHMETH,
       TREASURY_ADDRESS_V3,
@@ -594,7 +598,7 @@ describe("Pair Value", () => {
       toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
     ); // Balance for the staked Curve token
 
-    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, OHM_USD_RESERVE_BLOCK);
+    const records = getLiquidityBalances(TIMESTAMP, NATIVE_ETH, BLOCK_NUMBER);
 
     const totalValueExpected = getPairValue(
       ohmReserves,
@@ -620,7 +624,7 @@ describe("Pair Value", () => {
     mockFraxLockedBalanceZero();
     mockZeroWalletBalances(
       ERC20_BALANCER_OHM_DAI_WETH,
-      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID),
+      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID, BLOCK_NUMBER),
     );
 
     // Mock total value
@@ -647,10 +651,10 @@ describe("Pair Value", () => {
     );
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
-    mockZeroWalletBalances(ERC20_CVX_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
-    mockFraxLockedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
+    mockZeroWalletBalances(ERC20_CRV_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
+    mockFraxLockedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
     mockConvexStakedBalance(
       ERC20_CVX_FRAX_USDC,
       CONVEX_STAKING_PROXY_FRAXBP,
@@ -658,7 +662,7 @@ describe("Pair Value", () => {
       toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
     ); // Balance for the staked Curve token
 
-    const records = getLiquidityBalances(TIMESTAMP, ERC20_FRAX, OHM_USD_RESERVE_BLOCK);
+    const records = getLiquidityBalances(TIMESTAMP, ERC20_FRAX, BLOCK_NUMBER);
 
     const totalValueExpected = getPairValue(
       fraxReserves,
@@ -683,7 +687,7 @@ describe("Pair Value", () => {
     mockFraxLockedBalanceZero();
     mockZeroWalletBalances(
       ERC20_BALANCER_OHM_DAI_WETH,
-      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID),
+      getWalletAddressesForContract(POOL_BALANCER_OHM_DAI_WETH_ID, BLOCK_NUMBER),
     );
 
     // Mock price lookup
@@ -714,10 +718,10 @@ describe("Pair Value", () => {
     );
     // Mock balance
     const crvBalance = BigDecimal.fromString("10");
-    mockZeroWalletBalances(ERC20_CRV_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
-    mockZeroWalletBalances(ERC20_CVX_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
-    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH));
-    mockFraxLockedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC));
+    mockZeroWalletBalances(ERC20_CRV_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
+    mockZeroWalletBalances(ERC20_CVX_FRAX_USDC, getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
+    mockConvexStakedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_OHM_ETH, BLOCK_NUMBER));
+    mockFraxLockedBalanceZero(getWalletAddressesForContract(PAIR_CURVE_FRAX_USDC, BLOCK_NUMBER));
     mockFraxLockedBalance(
       ERC20_CVX_FRAX_USDC_STAKED,
       CONVEX_STAKING_PROXY_FRAXBP,
@@ -725,7 +729,7 @@ describe("Pair Value", () => {
       toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
     ); // Balance for the locked token
 
-    const records = getLiquidityBalances(TIMESTAMP, ERC20_FRAX, OHM_USD_RESERVE_BLOCK);
+    const records = getLiquidityBalances(TIMESTAMP, ERC20_FRAX, BLOCK_NUMBER);
 
     const totalValueExpected = getPairValue(
       fraxReserves,
@@ -749,7 +753,7 @@ describe("pair records", () => {
     // token1 OK
     mockERC20TotalSupply(ERC20_FRAX_BP, 18, BigInt.fromString("1000"));
 
-    const expectedRecords = getCurvePairRecords(TIMESTAMP, PAIR_CURVE_OHM_FRAXBP, null, BigInt.fromString("15000000"));
+    const expectedRecords = getCurvePairRecords(TIMESTAMP, PAIR_CURVE_OHM_FRAXBP, null, BLOCK_NUMBER);
 
     assert.i32Equals(0, expectedRecords.length);
   });
@@ -760,7 +764,7 @@ describe("pair records", () => {
     // Mock token1 reverting
     createMockedFunction(Address.fromString(ERC20_FRAX_BP), "decimals", "decimals():(uint8)").reverts();
 
-    const expectedRecords = getCurvePairRecords(TIMESTAMP, PAIR_CURVE_OHM_FRAXBP, null, BigInt.fromString("15000000"));
+    const expectedRecords = getCurvePairRecords(TIMESTAMP, PAIR_CURVE_OHM_FRAXBP, null, BLOCK_NUMBER);
 
     assert.i32Equals(0, expectedRecords.length);
   });