+++ subgraphs/shared/src/Wallets.ts
@@ -42,12 +42,9 @@ export const DAO_WORKING_CAPITAL = "0xF65A665D650B5De224F46D729e2bD0885EeA9dA5".
  */
 export const OLYMPUS_ASSOCIATION_WALLET = "0x4c71db02aeeb336cbd8f3d2cc866911f6e2fbd94".toLowerCase();
 
+// Don't iterate through these addresses: use getClearinghouseAddresses() instead
 export const COOLER_LOANS_CLEARINGHOUSE_V1 = "0xD6A6E8d9e82534bD65821142fcCd91ec9cF31880".toLowerCase();
 export const COOLER_LOANS_CLEARINGHOUSE_V1_1 = "0xE6343ad0675C9b8D3f32679ae6aDbA0766A2ab4c".toLowerCase();
-export const COOLER_LOANS_CLEARINGHOUSES = [
-  COOLER_LOANS_CLEARINGHOUSE_V1,
-  COOLER_LOANS_CLEARINGHOUSE_V1_1,
-];
 
 export const TRSRY = "0xa8687A15D4BE32CC8F0a8a7B9704a4C3993D9613".toLowerCase();
 export const TRSRY_V1_1 = "0xea1560F36F71a2F54deFA75ed9EaA15E8655bE22".toLowerCase();